+++ src/data/skill/public/init.lua
@@ -1,5 +1,5 @@
 return require 'data.load_file'("data.skill.public.", {
     "test",
     "test1",
-    "烈焰風暴",
+    "flamestrike",
 })