+++ src/war3/hero.lua
@@ -1,7 +1,4 @@
 local require = require
-local SkillTree = require 'lib.skill_tree.skill_tree'
-local skill_decorator = require 'lib.skill_decorator':new()
-local skill_manager = require 'lib.skill_manager':new()
 local Hero = require 'std.class'("Hero", require 'war3.unit')
 
 function Hero:_new(unit)
@@ -13,14 +10,14 @@ function Hero:__call(unit)
 end
 
 function Hero:decorateSkill(decorator_name)
-    skill_decorator:append(self, decorator_name)
+    require 'lib.skill_decorator':new():append(self, decorator_name)
     return self
 end
 
 -- target必須是Unit或其子類別
 function Hero:spell(skill_name, target, period)
-    local skill = skill_manager:get(skill_name, self, target)
-    return SkillTree:new(skill):append(skill.scripts):setPeriod(period):run()
+    local skill = require 'lib.skill_manager':new():get(skill_name, self, target)
+    return require 'lib.skill_tree.skill_tree':new(skill):append(skill.scripts):setPeriod(period):run()
 end
 
 return Hero