+++ src/war3/text.lua
@@ -15,6 +15,7 @@ local Init, SetTrace, Scaling, Move, MoveSin, ComputeSinTrace, ComputeMin, Updat
 --     font_size: {字體最小值, 字體增大值, 字體最大值},
 --     height: {基本高度, 增加高度(沒用填0), 最大高度},
 --     offset(可選): {距離, 角度},
+--     on_pulse(可選): 每個週期都會調用的函數,
 -- }
 function Text:_new(data)
     data._invalid_ = false
@@ -118,6 +119,11 @@ function Text:start()
         PERIOD,
         (self.time > 0) and self.time / PERIOD or -1,
         function(timer)
+            -- NOTE: 一定要放在update前面，不然on_pulse如果有更新文字之類的動作，本次週期不會更新。
+            if self.on_pulse then
+                self:on_pulse(timer)
+            end
+
             self:update(timer:getRuntime())
 
             if IsExpired(self, timer.count_) then
@@ -126,6 +132,8 @@ function Text:start()
             end
         end
     ):start()
+
+    return self
 end
 
 IsExpired = function(self, count)
@@ -134,6 +142,11 @@ end
 
 function Text:stop()
     self._invalid_ = true
+    return self
+end
+
+function Text:setText(text)
+    self.text = text
     return self
 end